+++ src/python/pants/backend/python/lint/black/register.py
@@ -6,9 +6,9 @@
 See https://black.readthedocs.io/en/stable/.
 """
 
-from pants.backend.python.lint import python_formatter
+from pants.backend.python.lint import python_fmt
 from pants.backend.python.lint.black import rules as black_rules
 
 
 def rules():
-    return (*black_rules.rules(), *python_formatter.rules())
+    return (*black_rules.rules(), *python_fmt.rules())
