+++ src/python/pants/backend/python/lint/isort/register.py
@@ -6,7 +6,7 @@
 See https://timothycrosley.github.io/isort/.
 """
 
-from pants.backend.python.lint import python_formatter
+from pants.backend.python.lint import python_fmt
 from pants.backend.python.lint.isort import rules as isort_rules
 from pants.backend.python.lint.isort.isort_prep import IsortPrep
 from pants.backend.python.lint.isort.isort_run import IsortRun
@@ -14,7 +14,7 @@ from pants.goal.task_registrar import TaskRegistrar as task
 
 
 def rules():
-    return (*isort_rules.rules(), *python_formatter.rules())
+    return (*isort_rules.rules(), *python_fmt.rules())
 
 
 def register_goals():
